feat: Update detect_reslt_list to be resuable
